Added an ItemTouchHelper,  OnStartDragListener and SimpleItemTouchHelperCallback.
These files do the following:
Itemtouchlistener - Is an interface class that will provide a list of methods that will eventually tie into callback methods in the ItemTouchHelper class.

OnStartDragListener - This will be called when the user begins a 'drag-and-drop' interaction with the touchscreen.

SimpleItemTouchHelperCallback - Makes sure that the recyclerview has the capability to receive and process gestures from the touchscreen