I have updated my saveditem activity
i added the ItemTouchHelper as a member variable to that we can use it in the OnStartDragListener's onStartDrag() method.

In the FirebaseRestaurantListAdapter constructor, this refers to the OnStartDragListener and the Context. Both are necessary to construct an instance of a FirebaseRestaurantListAdapter.

The SimpleItemTouchHelperCallback takes an adapter as a parameter so we pass it the instance of the FirebaseRestaurantListAdapter i had just created.

The ItemTouchHelper takes a ItemTouchHelper.Callback as an argument so i can pass it the instance of the SimpleItemTouchHelperCallback that i had just created using our adapter.

To enable the interfaces to communicate with the necessary callbacks, i must attach the ItemTouchHelper to our RecyclerView using the attachToRecyclerView() method.

Finally, i call the startDrag() method on the instance of our ItemTouchHelper inside of the onStartDrag() override which will eventually send our touch events back to our SimpleItemTouchHelperCallback.